+++ projects/ngx-bar-rating-demo/src/app/app.config.server.ts
@@ -4,8 +4,8 @@ import { appConfig } from './app.config';
 
 const serverConfig: ApplicationConfig = {
   providers: [
-    provideServerRendering()
+    provideServerRendering(),
   ]
 };
 
-export const config: ApplicationConfig = mergeApplicationConfig(appConfig, serverConfig);
+export const config = mergeApplicationConfig(appConfig, serverConfig);