+++ projects/ngx-bar-rating-demo/src/app/footer/footer.component.ts
@@ -2,10 +2,9 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { FaIconComponent } from '@fortawesome/angular-fontawesome';
 
 @Component({
-  standalone: true,
   selector: 'app-footer',
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss'],
+  styleUrl: './footer.component.scss',
   imports: [FaIconComponent],
   changeDetection: ChangeDetectionStrategy.OnPush
 })